# Conflicts:
#	src/main/java/at/qe/skeleton/models/SensorStation.java
#	src/main/java/at/qe/skeleton/models/Userx.java
#	src/main/java/at/qe/skeleton/models/demo/LogEntry.java
#	src/main/java/at/qe/skeleton/models/demo/LogEntryType.java
#	src/main/java/at/qe/skeleton/models/demo/Message.java
#	src/main/java/at/qe/skeleton/models/demo/UserStatus.java
#	src/main/java/at/qe/skeleton/models/demo/UserStatusInfo.java
#	src/main/java/at/qe/skeleton/services/UserService.java
#	src/test/java/at/qe/skeleton/tests/EqualsImplementationTest.java

+++ src/main/java/at/qe/skeleton/models/SensorValues.java
@@ -9,7 +9,7 @@ public class SensorValues {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "VALUES_ID")
-    private Long id;
+    private Integer id;
 
     @Column(name = "HUMIDITY")
     private Double humidity;
@@ -32,7 +32,7 @@ public class SensorValues {
     public SensorValues() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
