+++ src/main/java/at/qe/skeleton/repositories/MeasurementRepository.java
@@ -4,5 +4,5 @@ import at.qe.skeleton.models.Measurement;
 
 public interface MeasurementRepository extends AbstractRepository<Measurement, Long> {
 
-    Measurement findFirstById(Long id);
+    Measurement findFirstById(Integer id);
 }
